+++ client/package.json
@@ -1,0 +1,51 @@
+{
+  "name": "d2-extension-client",
+  "displayName": "D2",
+  "publisher": "Terrastruct",
+  "description": "Support for .d2 files.",
+  "version": "0.8.7",
+  "license": "BSD-3-Clause",
+  "contributors": [
+    "Barry Nolte <barry@barrynolte.com>"
+  ],
+  "engines": {
+    "vscode": "^1.73.0"
+  },
+  "devDependencies": {
+    "@types/glob": "^8.0.0",
+    "@types/markdown-it-container": "^2.0.5",
+    "@types/mocha": "^10.0.0",
+    "@types/node": "16.x",
+    "@types/vscode": "^1.73.0",
+    "@typescript-eslint/eslint-plugin": "^5.50.0",
+    "@typescript-eslint/parser": "^5.50.0",
+    "@vscode/test-electron": "^2.2.0",
+    "@vscode/vsce": "^2.15.0",
+    "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-formatter-pretty": "^4.1.0",
+    "eslint-webpack-plugin": "^3.2.0",
+    "glob": "^8.0.3",
+    "mocha": "^10.1.0",
+    "prettier": "^2.8.3",
+    "ts-loader": "^9.4.2",
+    "typescript": "^4.9.4",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0"
+  },
+  "dependencies": {
+    "markdown-it-container": "^3.0.0",
+    "path": "^0.12.7",
+    "vscode-languageclient": "^8.1.0"
+  },
+  "homepage": "https://d2lang.com",
+  "icon": "d2-icon.png",
+  "bugs": {
+    "url": "https://github.com/terrastruct/d2-vscode/issues",
+    "email": "info@terrastruct.com"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/terrastruct/d2-vscode"
+  }
+}
