+++ client/src/docToPreviewGenerator.ts
