+++ package.json
@@ -28,7 +28,7 @@
   "activationEvents": [
     "onStartupFinished"
   ],
-  "main": "./dist/extension.js",
+  "main": "./client/dist/extension.js",
   "contributes": {
     "taskDefinitions": [
       {
@@ -291,19 +291,20 @@
   },
   "scripts": {
     "vscode:prepublish": "yarn run package",
-    "compile": "webpack",
-    "watch": "webpack --watch --devtool source-map",
+    "compile": "tsc -b",
+    "watch": "tsc -b -w",
+    "lint": "eslint ./client/src ./server/src --ext .ts,.tsx",
+    "pretty": "npx prettier@2.8.1 --loglevel=warn --print-width=90 --write **/*.ts",
     "package": "webpack --mode production --devtool hidden-source-map",
+    "installdep": "cd client && npm install && cd ../server && npm install && cd ..",
     "dev": "code --uninstall-extension terrastruct.d2; yarn gen && yarn pkg && code --install-extension d2.vsix",
     "gen": "yq -o=json . syntaxes/d2.tmLanguage.yaml > syntaxes/d2.tmLanguage.json",
     "pkg": "vsce package --out d2.vsix"
   },
   "devDependencies": {
     "@types/glob": "^8.0.0",
-    "@types/markdown-it-container": "^2.0.5",
     "@types/mocha": "^10.0.0",
     "@types/node": "16.x",
-    "@types/vscode": "^1.73.0",
     "@typescript-eslint/eslint-plugin": "^5.50.0",
     "@typescript-eslint/parser": "^5.50.0",
     "@vscode/test-electron": "^2.2.0",
@@ -319,11 +320,6 @@
     "typescript": "^4.9.4",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0"
-  },
-  "dependencies": {
-    "markdown-it-container": "^3.0.0",
-    "path": "^0.12.7",
-    "vscode-languageclient": "^7.0.0"
   },
   "homepage": "https://d2lang.com",
   "icon": "d2-icon.png",
