--- src/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "ES2020",
-    "lib": ["ES2020"],
-    "esModuleInterop": false,
-    "sourceMap": true,
-    "rootDir": "../src",
-    "strict": true
-  }
-}
